Semester_2/СollectionAndProcessingInformation/parser/pyproject.toml
[tool.poetry]
name = "parser"
version = "0.1.0"
description = ""
authors = ["DmitryMogilnikov <d.mogqs@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pdfminer-six = "^20221105"
flake8 = "^6.0.0"
isort = "^5.12.0"
pytest = "^7.2.1"
python-docx = "^0.8.11"
docx = "^0.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
